Refactor field guesser
