+++ Resources/config/doctrine_odm.xml
@@ -7,14 +7,14 @@
         <parameter key="admingenerator.doctrine_odm.class">Admingenerator\GeneratorBundle\Generator\DoctrineODMGenerator</parameter>
         <parameter key="admingenerator.fieldguesser.doctrine_odm.class">Admingenerator\GeneratorBundle\Guesser\DoctrineODMFieldGuesser</parameter>
         <parameter key="admingenerator.queryfilter.doctrine_odm.class">Admingenerator\GeneratorBundle\QueryFilter\DoctrineODMQueryFilter</parameter>
-        <parameter key="admingenerator.sessionserializer.doctrine_odm.class">Admingenerator\GeneratorBundle\SessionSerializer\DoctrineODMSessionSerializer</parameter>
     </parameters>
     
     <services>
         <!-- Doctine MongoDB -->
         <service id="admingenerator.fieldguesser.doctrine_odm" class="%admingenerator.fieldguesser.doctrine_odm.class%"
             public="false">
-            <argument type="service" id="doctrine.odm.mongodb.document_manager"></argument>
+            <argument type="service" id="doctrine"></argument>
+            <argument type="string">document</argument>
         </service>
 
         <service id="admingenerator.generator.doctrine_odm"
@@ -37,10 +37,5 @@
                 <argument type="service" id="kernel" />
             </call>
         </service>
-
-        <service id="admingenerator.sessionserializer.doctrine_odm" class="%admingenerator.sessionserializer.doctrine_odm.class%">
-            <argument type="service" id="doctrine.odm.mongodb.document_manager"></argument>
-        </service>
-        
     </services>
 </container>
